WIP - JoinKeysignDoneSummary
