+++ VultisigApp/VultisigApp/Views/Keysign/JoinKeysignDoneView.swift
@@ -12,7 +12,6 @@ struct JoinKeysignDoneView: View {
     @ObservedObject var viewModel: KeysignViewModel
     @Binding var showAlert: Bool
     
-    @Environment(\.openURL) var openURL
     @Environment(\.dismiss) var dismiss
     
     var body: some View {
@@ -29,88 +28,7 @@ struct JoinKeysignDoneView: View {
     }
     
     var cards: some View {
-        ScrollView {
-            if viewModel.txid.isEmpty {
-                transactionComplete
-            } else {
-                if let approveTxid = viewModel.approveTxid {
-                    card(title: NSLocalizedString("Approve", comment: ""), txid: approveTxid)
-                }
-
-                card(title: NSLocalizedString("transaction", comment: "Transaction"), txid: viewModel.txid)
-            }
-        }
-    }
-
-    func card(title: String, txid: String) -> some View {
-        VStack(alignment: .leading, spacing: 8) {
-            titleSection(title: title, txid: txid)
-
-            Text(txid)
-                .font(.body13Menlo)
-                .foregroundColor(.turquoise600)
-
-            if viewModel.txid == txid, let link = viewModel.getSwapProgressURL(txid: viewModel.txid) {
-                HStack {
-                    Spacer()
-                    progressButton(link: link)
-                }
-            }
-        }
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .padding(16)
-        .background(Color.blue600)
-        .cornerRadius(10)
-        .padding(.horizontal, 16)
-    }
-    
-    var transactionComplete: some View {
-        Text(NSLocalizedString("transactionComplete", comment: "Transaction"))
-            .font(.body24MontserratMedium)
-            .foregroundColor(.neutral0)
-    }
-    
-    func titleSection(title: String, txid: String) -> some View {
-        HStack(spacing: 12) {
-            Text(title)
-                .font(.body20MontserratSemiBold)
-                .foregroundColor(.neutral0)
-            
-            copyButton(txid: txid)
-            linkButton(txid: txid)
-        }
-    }
-    
-    func copyButton(txid: String) -> some View {
-        Button {
-            copyHash(txid: txid)
-        } label: {
-            Image(systemName: "square.on.square")
-                .font(.body18Menlo)
-                .foregroundColor(.neutral0)
-        }
-        
-    }
-    
-    func linkButton(txid: String) -> some View {
-        Button {
-            shareLink(txid: txid)
-        } label: {
-            Image(systemName: "link")
-                .font(.body18Menlo)
-                .foregroundColor(.neutral0)
-        }
-        
-    }
-
-    func progressButton(link: String) -> some View {
-        Button {
-            progressLink(link: link)
-        } label: {
-            Text(NSLocalizedString("Swap progress", comment: ""))
-                .font(.body14Menlo)
-                .foregroundColor(.neutral0)
-        }
+        JoinKeysignDoneSummary(viewModel: viewModel, showAlert: $showAlert)
     }
 
     var continueButton: some View {
@@ -128,19 +46,6 @@ struct JoinKeysignDoneView: View {
             .font(.body)
             .bold()
             .foregroundColor(.neutral0)
-    }
-    
-    private func shareLink(txid: String) {
-        let urlString = viewModel.getTransactionExplorerURL(txid: txid)
-        if !urlString.isEmpty, let url = URL(string: urlString) {
-            openURL(url)
-        }
-    }
-
-    private func progressLink(link: String) {
-        if !link.isEmpty, let url = URL(string: link) {
-            openURL(url)
-        }
     }
 }
