+++ VultisigApp/VultisigApp/iOS/View/Keysign/JoinKeysignDoneView+iOS.swift
@@ -8,7 +8,7 @@
 #if os(iOS)
 import SwiftUI
 
-extension JoinKeysignDoneView {
+extension JoinKeysignDoneSummary {
     func copyHash(txid: String) {
         let urlStr = viewModel.getTransactionExplorerURL(txid: txid)
         showAlert = true
